# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Auth Service API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthTokens {
  accessToken: string;
  refreshToken: string;
  expiresIn: number;
}

export interface RegisterInput {
  projectCode: string;
  email: string;
  /** @minLength 8 */
  password: string;
  fullName: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginInput {
  projectCode: string;
  email: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface ForgotPasswordInput {
  projectCode: string;
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  /** @minLength 8 */
  password: string;
}

export type VerifyEmailParams = {
token: string;
};