# lib/api-zod/src/generated/types/authTokens.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Auth Service API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthTokens {
  accessToken: string;
  refreshToken: string;
  expiresIn: number;
}