# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Auth Service API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  projectCode: string;
  email: string;
  password: string;
}