# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Auth Service API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  projectCode: string;
  email: string;
  /** @minLength 8 */
  password: string;
  fullName: string;
  /** @nullable */
  phone?: string | null;
}