# lib/api-zod/src/generated/types/resetPasswordInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Auth Service API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordInput {
  token: string;
  /** @minLength 8 */
  password: string;
}